backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==2.2.0
scipy==1.15.3
pandas==2.2.3
python-dotenv
